king-ai-v3/agentic-framework-main/tools/kautilya/requirements.txt
# CLI Core
click>=8.1.7
rich>=13.7.0
prompt-toolkit>=3.0.43
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
httpx>=0.26.0
jinja2>=3.1.2
questionary>=2.0.1
anyio>=4.2.0
jsonschema>=4.20.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Web Search
ddgs>=9.0.0
beautifulsoup4>=4.12.0

# Document Processing
python-docx>=1.1.0
pypdf>=4.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
markitdown>=0.1.0
Pillow>=10.0.0

# RAG Pipeline
haystack-ai>=2.0.0
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
